arm: psci: correct register type
Use unsigned long for regs. Using u32 will cut the value on AArch64.

+++ hypervisor/arch/arm-common/psci.c
@@ -70,7 +70,7 @@ static long psci_emulate_affinity_info(struct per_cpu *cpu_data,
 long psci_dispatch(struct trap_context *ctx)
 {
 	struct per_cpu *cpu_data = this_cpu_data();
-	u32 function_id = ctx->regs[0];
+	unsigned long function_id = ctx->regs[0];
 
 	this_cpu_data()->stats[JAILHOUSE_CPU_STAT_VMEXITS_PSCI]++;
